+++ Changelog.md
@@ -10,6 +10,7 @@
 - Changed the product stock field to be cast to float
 - Removed the Buyable interface/trait from `Foundation\MasterProduct` - it was conceptually wrong;
 - Added `Buyable` to `Foundation\MasterProductVariant`
+- Added the `Product::findBySku()` method to the base product class
 
 ## 3.2.0
 ##### 2022-12-08